sanxei95 - * refactor app.tsx
