+++ src/features/views/components/views-page/views-page.tsx
@@ -1,4 +1,5 @@
 import { useState } from "react";
+import { useNavigate, useLocation } from "react-router-dom";
 import { Plus, User, LogOut } from "lucide-react";
 import {
   DropdownMenu,
@@ -6,56 +7,62 @@ import {
   DropdownMenuItem,
   DropdownMenuTrigger,
 } from "@radix-ui/react-dropdown-menu";
+import { useAppSelector } from "@/app/hooks.ts";
+import { selectIsAuthenticated, selectUsername } from "@/app/authSlice.ts";
+import { logout } from "@/features/auth/authApi.ts";
+import { useViewsData } from "@/features/views/hooks/useViewsData.ts";
 import { View } from "@/features/views/model/view.ts";
 import { ViewsList } from "./views-list.tsx";
 import { CreateView } from "./actions/create-view.tsx";
 import "./views-page.css";
 
-interface ViewsPageProps {
-  views: View[];
-  isLoadingViews: boolean;
-  isAuthenticated: boolean;
-  username: string | null;
-  activeTab: "featured" | "own";
-  onTabChange: (tab: "featured" | "own") => void;
-  onViewClick: (viewId: string) => void;
-  onCreateView: (pendingView: View) => void;
-  onDeleteView: (viewId: string) => void;
-  onLoginRequired: () => void;
-  onLogout: () => void;
-}
+export function ViewsPage() {
+  const navigate = useNavigate();
+  const location = useLocation();
+  const isAuthenticated = useAppSelector(selectIsAuthenticated);
+  const username = useAppSelector(selectUsername);
 
-export function ViewsPage({
-  views,
-  isLoadingViews,
-  isAuthenticated,
-  username,
-  activeTab,
-  onTabChange,
-  onViewClick,
-  onCreateView,
-  onDeleteView,
-  onLoginRequired,
-  onLogout,
-}: Readonly<ViewsPageProps>) {
+  const [activeTab, setActiveTab] = useState<"featured" | "own">(
+    isAuthenticated ? "own" : "featured",
+  );
   const [isCreateDialogOpen, setIsCreateDialogOpen] = useState(false);
 
+  const { featuredViews, isLoadingFeatured, ownViews, isLoadingOwn, createView, deleteView } =
+    useViewsData(isAuthenticated);
+
+  const views = activeTab === "featured" ? featuredViews : ownViews;
+  const isLoadingViews = activeTab === "featured" ? isLoadingFeatured : isLoadingOwn;
   const isSyncing = views.some((v) => !v.isSynced);
 
-  const handleCreateClick = () => {
+  const requireAuth = (action: () => void) => {
     if (!isAuthenticated) {
-      onLoginRequired();
+      navigate("/login", { state: { from: location.pathname } });
       return;
     }
-    setIsCreateDialogOpen(true);
+    action();
   };
 
-  const handleOwnTabClick = () => {
-    if (!isAuthenticated) {
-      onLoginRequired();
-      return;
-    }
-    onTabChange("own");
+  const handleViewClick = (viewId: string) => {
+    const view = views.find((v) => v.id === viewId);
+    navigate(`/${viewId}`, {
+      state: {
+        owner: view?.simpleView.owner,
+        entitiesCount: view?.simpleView.entitiesIds.length ?? 0,
+      },
+    });
+  };
+
+  const handleCreateClick = () => requireAuth(() => setIsCreateDialogOpen(true));
+
+  const handleOwnTabClick = () => requireAuth(() => setActiveTab("own"));
+
+  const handleCreateView = (pendingView: View) => createView(pendingView);
+
+  const handleDeleteView = (viewId: string) => requireAuth(() => deleteView(viewId));
+
+  const handleLogout = async () => {
+    await logout();
+    navigate("/");
   };
 
   return (
@@ -88,7 +95,7 @@ export function ViewsPage({
                 <DropdownMenuContent className="user-menu-content" align="end">
                   <DropdownMenuItem
                     className="user-menu-item user-menu-item--danger"
-                    onSelect={onLogout}
+                    onSelect={handleLogout}
                   >
                     <LogOut className="user-menu-item-icon" />
                     Logout
@@ -108,7 +115,7 @@ export function ViewsPage({
           <div className="views-tab-toggle">
             <button
               className={`views-tab-btn${activeTab === "featured" ? " views-tab-btn--active" : ""}`}
-              onClick={() => onTabChange("featured")}
+              onClick={() => setActiveTab("featured")}
             >
               Featured ladders
             </button>
@@ -125,16 +132,16 @@ export function ViewsPage({
           views={views}
           isLoadingViews={isLoadingViews}
           username={username}
-          onViewClick={onViewClick}
+          onViewClick={handleViewClick}
           onCreateView={handleCreateClick}
-          onDeleteView={onDeleteView}
+          onDeleteView={handleDeleteView}
         />
       </div>
 
       <CreateView
         open={isCreateDialogOpen}
         onOpenChange={setIsCreateDialogOpen}
-        onCreateView={onCreateView}
+        onCreateView={handleCreateView}
       />
     </div>
   );
